+++ app/src/main/java/com/scooter/datacollector/domain/models/Session.kt
@@ -1,7 +1,7 @@
 package com.scooter.datacollector.domain.models
 
 data class Session(
-    val id: Int,
-    val usedId: Int,
+    val id: Long,
+    val usedId: Long,
     val rideMode: RideMode,
 )
