+++ app/src/main/java/com/scooter/datacollector/domain/repositories/ISessionRepository.kt
@@ -3,6 +3,6 @@ package com.scooter.datacollector.domain.repositories
 import com.scooter.datacollector.domain.models.Session
 
 interface ISessionRepository {
-    public fun getIdForNewSession(): Int
+    public fun getIdForNewSession(): Long
     public fun createSession(session: Session)
 }
